pyjokes/source/requirements.txt
tensorflow == 2.3.1
psycopg2-binary == 2.8.6